+++ spec/features/staff/users_can_edit_a_fund_spec.rb
@@ -1,23 +1,20 @@
 RSpec.feature "Users can edit a fund" do
-  before do
-    authenticate!(user: user)
-  end
-
-  let(:fund) { create(:fund, name: "old name", organisation: organisation) }
-  let(:organisation) { create(:organisation, name: "UKSA") }
-  let(:user) { create(:user, organisations: [organisation]) }
-
   context "when the user is not logged in" do
     it "redirects the user to the root path" do
+      organisation = create(:organisation)
+      fund = create(:fund, organisation: organisation)
       page.set_rack_session(userinfo: nil)
       visit edit_organisation_fund_path(organisation, fund)
       expect(current_path).to eq(root_path)
     end
   end
 
-  context "when no activity record exists" do
-    scenario "successfully editing a fund" do
+  context "when the user is an administrator" do
+    scenario "the user can edit a fund" do
+      organisation = create(:organisation)
       fund = create(:fund, organisation: organisation)
+      user = create(:administrator)
+      authenticate!(user: user)
 
       visit dashboard_path
       click_on(I18n.t("page_content.dashboard.button.manage_organisations"))
@@ -35,10 +32,12 @@ RSpec.feature "Users can edit a fund" do
     end
   end
 
-  context "when an activity record exists" do
-    scenario "successfully editing a fund" do
+  context "when the user is a delivery partner" do
+    scenario "the user cannot edit a fund" do
+      organisation = create(:organisation)
       fund = create(:fund, organisation: organisation)
-      create(:activity, hierarchy: fund)
+      user = create(:delivery_partner, organisations: [organisation])
+      authenticate!(user: user)
 
       visit dashboard_path
       click_on(I18n.t("page_content.dashboard.button.manage_organisations"))
@@ -46,13 +45,7 @@ RSpec.feature "Users can edit a fund" do
       click_on(organisation.name)
       click_on(fund.name)
 
-      within(".fund_name") do
-        click_on(I18n.t("generic.link.edit"))
-      end
-
-      fill_in "fund[name]", with: "My New Fund name"
-      click_button I18n.t("generic.button.submit")
-      expect(page).to have_content("My New Fund name")
+      expect(page).not_to have_content(I18n.t("generic.link.edit"))
     end
   end
 end